- ipv6 patch support - bug fixing IpcSocket.cpp, if length of buffer or string is zero no read or write operation

+++ IpcSocket.h
@@ -57,7 +57,7 @@ public:
 	
 	void send(int);
 	
-	void send(Octet *, unsigned int);
+	void send(Octet *, ssize_t);
 	
 	int recvInt(void);
